# openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "Matchbook API",
    "description": "Self-hosted agency finance ops \u2014 invoice collection, reconciliation, reporting.",
    "version": "0.1.0"
  },
  "paths": {
...
    "/vendors": {
      "get": {
        "tags": [
          "vendors",
          "vendors"
        ],
        "summary": "List Vendors",
        "operationId": "list_vendors_vendors_get",
        "parameters": [
          {
            "name": "skip",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "minimum": 0,
              "default": 0,
              "title": "Skip"
            }
          },
          {
            "name": "limit",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "maximum": 100,
              "minimum": 1,
              "default": 50,
              "title": "Limit"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/VendorList"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      },
      "post": {
        "tags": [
          "vendors",
          "vendors"
        ],
        "summary": "Create Vendor",
        "operationId": "create_vendor_vendors_post",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/VendorCreate"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/VendorRead"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/vendors/{vendor_id}": {
      "get": {
        "tags": [
          "vendors",
          "vendors"
        ],
        "summary": "Get Vendor",
        "operationId": "get_vendor_vendors__vendor_id__get",
        "parameters": [
          {
            "name": "vendor_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid",
              "title": "Vendor Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/VendorRead"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      },
      "patch": {
        "tags": [
          "vendors",
          "vendors"
        ],
        "summary": "Update Vendor",
        "operationId": "update_vendor_vendors__vendor_id__patch",
        "parameters": [
          {
            "name": "vendor_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid",
              "title": "Vendor Id"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/VendorUpdate"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/VendorRead"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      },
      "delete": {
        "tags": [
          "vendors",
          "vendors"
        ],
        "summary": "Delete Vendor",
        "operationId": "delete_vendor_vendors__vendor_id__delete",
        "parameters": [
          {
            "name": "vendor_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid",
              "title": "Vendor Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/VendorRead"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/vendors/{vendor_id}/invoices": {
      "get": {
        "tags": [
          "vendors",
          "vendors"
        ],
        "summary": "Get Vendor Invoices",
        "description": "Get invoice history for a vendor.",
        "operationId": "get_vendor_invoices_vendors__vendor_id__invoices_get",
        "parameters": [
          {
            "name": "vendor_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid",
              "title": "Vendor Id"
            }
          },
          {
            "name": "period",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Period"
            }
          },
          {
            "name": "skip",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "minimum": 0,
              "default": 0,
              "title": "Skip"
            }
          },
          {
            "name": "limit",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "maximum": 100,
              "minimum": 1,
              "default": 50,
              "title": "Limit"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/InvoiceList"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
    "/reports/generate": {
      "post": {
        "tags": [
          "reports",
          "reports"
        ],
        "summary": "Generate Report",
        "description": "Generate a financial report in PDF or Excel format and upload to Drive.",
        "operationId": "generate_report_reports_generate_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ReportRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful Response"
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/reports/preview": {
      "post": {
        "tags": [
          "reports",
          "reports"
        ],
        "summary": "Preview Report",
        "description": "Preview what a report would contain without generating it.",
        "operationId": "preview_report_reports_preview_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ReportRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ReportMeta"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
      "BillingCycle": {
        "type": "string",
        "enum": [
          "monthly",
          "annual",
          "one_off"
        ],
        "title": "BillingCycle"
      },
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "type": "array",
            "title": "Detail"
          }
        },
        "type": "object",
        "title": "HTTPValidationError"
      },
...
      "InvoiceList": {
        "properties": {
          "items": {
            "items": {
              "$ref": "#/components/schemas/InvoiceRead"
            },
            "type": "array",
            "title": "Items"
          },
          "total": {
            "type": "integer",
            "title": "Total"
          }
        },
        "type": "object",
        "required": [
          "items",
          "total"
        ],
        "title": "InvoiceList"
      },
      "InvoiceRead": {
        "properties": {
          "id": {
            "type": "string",
            "format": "uuid",
            "title": "Id"
          },
          "vendor": {
            "type": "string",
            "title": "Vendor"
          },
          "amount_excl": {
            "type": "string",
            "pattern": "^(?!^[-+.]*$)[+-]?0*\\d*\\.?\\d*$",
            "title": "Amount Excl"
          },
          "amount_incl": {
            "type": "string",
            "pattern": "^(?!^[-+.]*$)[+-]?0*\\d*\\.?\\d*$",
            "title": "Amount Incl"
          },
          "vat_amount": {
            "type": "string",
            "pattern": "^(?!^[-+.]*$)[+-]?0*\\d*\\.?\\d*$",
            "title": "Vat Amount"
          },
          "vat_rate": {
            "type": "string",
            "pattern": "^(?!^[-+.]*$)[+-]?0*\\d*\\.?\\d*$",
            "title": "Vat Rate"
          },
          "invoice_date": {
            "type": "string",
            "format": "date",
            "title": "Invoice Date"
          },
          "invoice_number": {
            "type": "string",
            "title": "Invoice Number"
          },
          "currency": {
            "type": "string",
            "title": "Currency"
          },
          "category": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Category"
          },
          "drive_url": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Drive Url"
          },
          "drive_file_id": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Drive File Id"
          },
          "source": {
            "$ref": "#/components/schemas/InvoiceSource"
          },
          "status": {
            "$ref": "#/components/schemas/InvoiceStatus"
          },
          "period": {
            "type": "string",
            "title": "Period"
          },
          "raw_extraction": {
            "anyOf": [
              {
                "additionalProperties": true,
                "type": "object"
              },
              {
                "type": "null"
              }
            ],
            "title": "Raw Extraction"
          },
          "created_at": {
            "type": "string",
            "format": "date-time",
            "title": "Created At"
          }
        },
        "type": "object",
        "required": [
          "id",
          "vendor",
          "amount_excl",
          "amount_incl",
          "vat_amount",
          "vat_rate",
          "invoice_date",
          "invoice_number",
          "currency",
          "source",
          "status",
          "period",
          "created_at"
        ],
        "title": "InvoiceRead"
      },
      "InvoiceSource": {
        "type": "string",
        "enum": [
          "gmail",
          "portal",
          "manual"
        ],
        "title": "InvoiceSource"
      },
      "InvoiceStatus": {
        "type": "string",
        "enum": [
          "pending",
          "matched",
          "unmatched",
          "flagged"
        ],
        "title": "InvoiceStatus"
      },
...
      "ReportFormat": {
        "type": "string",
        "enum": [
          "pdf",
          "excel"
        ],
        "title": "ReportFormat"
      },
      "ReportMeta": {
        "properties": {
          "timeframe_label": {
            "type": "string",
            "title": "Timeframe Label"
          },
          "periods": {
            "items": {
              "type": "string"
            },
            "type": "array",
            "title": "Periods"
          },
          "format": {
            "$ref": "#/components/schemas/ReportFormat"
          },
          "filename": {
            "type": "string",
            "title": "Filename"
          }
        },
        "type": "object",
        "required": [
          "timeframe_label",
          "periods",
          "format",
          "filename"
        ],
        "title": "ReportMeta"
      },
      "ReportRequest": {
        "properties": {
          "timeframe": {
            "$ref": "#/components/schemas/TimeframeType"
          },
          "format": {
            "$ref": "#/components/schemas/ReportFormat"
          },
          "period": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Period"
          },
          "quarter": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Quarter"
          },
          "year": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Year"
          },
          "from_period": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "From Period"
          },
          "to_period": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "To Period"
          }
        },
        "type": "object",
        "required": [
          "timeframe",
          "format"
        ],
        "title": "ReportRequest"
      },
...
      "TimeframeType": {
        "type": "string",
        "enum": [
          "single_month",
          "quarter",
          "ytd",
          "full_year",
          "custom"
        ],
        "title": "TimeframeType"
      },
...
      "ValidationError": {
        "properties": {
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "type": "array",
            "title": "Location"
          },
          "msg": {
            "type": "string",
            "title": "Message"
          },
          "type": {
            "type": "string",
            "title": "Error Type"
          },
          "input": {
            "title": "Input"
          },
          "ctx": {
            "type": "object",
            "title": "Context"
          }
        },
        "type": "object",
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError"
      },
      "VendorCreate": {
        "properties": {
          "name": {
            "type": "string",
            "title": "Name"
          },
          "aliases": {
            "anyOf": [
              {
                "items": {
                  "type": "string"
                },
                "type": "array"
              },
              {
                "type": "null"
              }
            ],
            "title": "Aliases"
          },
          "default_category": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Default Category"
          },
          "default_vat_rate": {
            "anyOf": [
              {
                "type": "number"
              },
              {
                "type": "string",
                "pattern": "^(?!^[-+.]*$)[+-]?0*\\d*\\.?\\d*$"
              },
              {
                "type": "null"
              }
            ],
            "title": "Default Vat Rate"
          },
          "billing_cycle": {
            "$ref": "#/components/schemas/BillingCycle",
            "default": "monthly"
          }
        },
        "type": "object",
        "required": [
          "name"
        ],
        "title": "VendorCreate"
      },
      "VendorList": {
        "properties": {
          "items": {
            "items": {
              "$ref": "#/components/schemas/VendorRead"
            },
            "type": "array",
            "title": "Items"
          },
          "total": {
            "type": "integer",
            "title": "Total"
          }
        },
        "type": "object",
        "required": [
          "items",
          "total"
        ],
        "title": "VendorList"
      },
      "VendorRead": {
        "properties": {
          "id": {
            "type": "string",
            "format": "uuid",
            "title": "Id"
          },
          "name": {
            "type": "string",
            "title": "Name"
          },
          "aliases": {
            "anyOf": [
              {
                "items": {
                  "type": "string"
                },
                "type": "array"
              },
              {
                "type": "null"
              }
            ],
            "title": "Aliases"
          },
          "default_category": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Default Category"
          },
          "default_vat_rate": {
            "anyOf": [
              {
                "type": "string",
                "pattern": "^(?!^[-+.]*$)[+-]?0*\\d*\\.?\\d*$"
              },
              {
                "type": "null"
              }
            ],
            "title": "Default Vat Rate"
          },
          "billing_cycle": {
            "$ref": "#/components/schemas/BillingCycle"
          }
        },
        "type": "object",
        "required": [
          "id",
          "name",
          "billing_cycle"
        ],
        "title": "VendorRead"
      },
...
      "VendorUpdate": {
        "properties": {
          "name": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Name"
          },
          "aliases": {
            "anyOf": [
              {
                "items": {
                  "type": "string"
                },
                "type": "array"
              },
              {
                "type": "null"
              }
            ],
            "title": "Aliases"
          },
          "default_category": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Default Category"
          },
          "default_vat_rate": {
            "anyOf": [
              {
                "type": "number"
              },
              {
                "type": "string",
                "pattern": "^(?!^[-+.]*$)[+-]?0*\\d*\\.?\\d*$"
              },
              {
                "type": "null"
              }
            ],
            "title": "Default Vat Rate"
          },
          "billing_cycle": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/BillingCycle"
              },
              {
                "type": "null"
              }
            ]
          }
        },
        "type": "object",
        "title": "VendorUpdate"
      },
...
    }
  }
}